LayoutTests/imported/w3c/web-platform-tests/dom/nodes/Document-createEvent-expected.txt


PASS CustomEvent should be an alias for CustomEvent. 
FAIL createEvent('CustomEvent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS customevent should be an alias for CustomEvent. 
FAIL createEvent('customevent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS CUSTOMEVENT should be an alias for CustomEvent. 
FAIL createEvent('CUSTOMEVENT') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS Event should be an alias for Event. 
FAIL createEvent('Event') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS event should be an alias for Event. 
FAIL createEvent('event') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS EVENT should be an alias for Event. 
FAIL createEvent('EVENT') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS Events should be an alias for Event. 
FAIL createEvent('Events') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS events should be an alias for Event. 
FAIL createEvent('events') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS EVENTS should be an alias for Event. 
FAIL createEvent('EVENTS') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS HTMLEvents should be an alias for Event. 
FAIL createEvent('HTMLEvents') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS htmlevents should be an alias for Event. 
FAIL createEvent('htmlevents') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS HTMLEVENTS should be an alias for Event. 
FAIL createEvent('HTMLEVENTS') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS KeyboardEvent should be an alias for KeyboardEvent. 
FAIL createEvent('KeyboardEvent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS keyboardevent should be an alias for KeyboardEvent. 
FAIL createEvent('keyboardevent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS KEYBOARDEVENT should be an alias for KeyboardEvent. 
FAIL createEvent('KEYBOARDEVENT') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
FAIL KeyEvents should be an alias for KeyboardEvent. NotSupportedError: DOM Exception 9
FAIL createEvent('KeyEvents') should be initialized correctly. undefined is not an object (evaluating 'ev.type')
FAIL keyevents should be an alias for KeyboardEvent. NotSupportedError: DOM Exception 9
FAIL createEvent('keyevents') should be initialized correctly. undefined is not an object (evaluating 'ev.type')
FAIL KEYEVENTS should be an alias for KeyboardEvent. NotSupportedError: DOM Exception 9
FAIL createEvent('KEYEVENTS') should be initialized correctly. undefined is not an object (evaluating 'ev.type')
PASS MessageEvent should be an alias for MessageEvent. 
FAIL createEvent('MessageEvent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS messageevent should be an alias for MessageEvent. 
FAIL createEvent('messageevent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS MESSAGEEVENT should be an alias for MessageEvent. 
FAIL createEvent('MESSAGEEVENT') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS MouseEvent should be an alias for MouseEvent. 
FAIL createEvent('MouseEvent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS mouseevent should be an alias for MouseEvent. 
FAIL createEvent('mouseevent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS MOUSEEVENT should be an alias for MouseEvent. 
FAIL createEvent('MOUSEEVENT') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS MouseEvents should be an alias for MouseEvent. 
FAIL createEvent('MouseEvents') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS mouseevents should be an alias for MouseEvent. 
FAIL createEvent('mouseevents') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS MOUSEEVENTS should be an alias for MouseEvent. 
FAIL createEvent('MOUSEEVENTS') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
FAIL TouchEvent should be an alias for TouchEvent. NotSupportedError: DOM Exception 9
FAIL createEvent('TouchEvent') should be initialized correctly. undefined is not an object (evaluating 'ev.type')
FAIL touchevent should be an alias for TouchEvent. NotSupportedError: DOM Exception 9
FAIL createEvent('touchevent') should be initialized correctly. undefined is not an object (evaluating 'ev.type')
FAIL TOUCHEVENT should be an alias for TouchEvent. NotSupportedError: DOM Exception 9
FAIL createEvent('TOUCHEVENT') should be initialized correctly. undefined is not an object (evaluating 'ev.type')
PASS UIEvent should be an alias for UIEvent. 
FAIL createEvent('UIEvent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS uievent should be an alias for UIEvent. 
FAIL createEvent('uievent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS UIEVENT should be an alias for UIEvent. 
FAIL createEvent('UIEVENT') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS UIEvents should be an alias for UIEvent. 
FAIL createEvent('UIEvents') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS uievents should be an alias for UIEvent. 
FAIL createEvent('uievents') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS UIEVENTS should be an alias for UIEvent. 
FAIL createEvent('UIEVENTS') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined
PASS Should throw NOT_SUPPORTED_ERR for unrecognized arguments 
